[Uday] Add. Section2 uc2 Introducing service layer.

+++ src/main/java/com/payroll/app/model/Employee.java
@@ -6,23 +6,23 @@ import jakarta.persistence.*;
 public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private double salary;
 
     public Employee(){}
 
-    public Employee(Integer id, String name, double salary) {
+    public Employee(Long id, String name, double salary) {
         this.id = id;
         this.name = name;
         this.salary = salary;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
